now everything compiles properly: corrected unit_tests for missing adaptive subcells. added methods for celltraversal for linkedcells (not yet implemented)
git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/branches/overlappingComm@3385 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/parallel/NonBlockingMPIMultiStepHandler.cpp
@@ -7,6 +7,11 @@
 
 #include "NonBlockingMPIMultiStepHandler.h"
 #include <assert.h>
+#include "utils/Logger.h"
+#include "utils/Timer.h"
+#include "DomainDecompMPIBase.h"
+#include "particleContainer/ParticleContainer.h"
+
 
 using Log::global_log;
 using namespace std;
@@ -31,7 +36,7 @@ void NonBlockingMPIMultiStepHandler::performComputation() {
 
 	assert(stageCount > 0);
 
-	for (unsigned int i = 0; i < stageCount; ++i) {
+	for (unsigned int i = 0; i < static_cast<unsigned int>(stageCount); ++i) {
 		_decompositionTimer->start();
 		_domainDecomposition->prepareNonBlockingStage(false,
 				_moleculeContainer, _domain, i);
